% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat2bbox.R
\name{lonlat2bbox}
\alias{lonlat2bbox}
\title{Bounding box from lon and lat values}
\usage{
lonlat2bbox(lon, lat, width = NULL, dist = NULL, units = "kilometers",
  method = "sp_rgeos", ...)
}
\arguments{
\item{lon}{(numeric) longitude}

\item{lat}{(numeric) latitude}

\item{width}{(numeric) width, passed on to [rgeos::gBuffer()], see it's 
docs for more info. units are metric, so 10 is 10 m, 1000 is 1000 m, 
etc.}

\item{dist}{(numeric) distance. passed on to [lawn::lawn_buffer()]}

\item{units}{(character) units. passed on to [lawn::lawn_buffer()]. 
Default: kilometers}

\item{method}{(character) which method to use, either combination of \pkg{sp}
and \pkg{rgeos} ("sp_rgeos") or \pkg{lawn}}

\item{...}{additional parameters passed on to [rgeos::gBuffer()]}
}
\value{
bounding box numeric values in the order `[minX, minY, maxX, maxY]`
}
\description{
Bounding box from lon and lat values
}
\examples{
# 10m
lonlat2bbox(lon=-120, lat=45, width=10)
# 100m
lonlat2bbox(-120, 45, 100)
# 1000m, or 1 km
lonlat2bbox(-120, 45, 1000)
# 10 km
lonlat2bbox(-120, 45, 10^4)
}
